requirements.txt
# EST Backend Requirements
# Python 3.10+

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Configuration
pydantic==2.5.0
pydantic-settings==2.1.0

# Authentication & Security
bcrypt==4.1.1
PyJWT==2.8.1

# HTTP Client (for inter-service communication)
httpx==0.25.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
